Denies the admin user to delete himself !

+++ src/Controller/UsersController.php
@@ -177,13 +177,22 @@ class UsersController extends AppController
     {
         $this->request->allowMethod(['post', 'delete']);
         $user = $this->Users->get($id);
-        if ($this->Users->delete($user)) {
-            $this->Flash->success(__('The user has been deleted.'));
-        } else {
-            $this->Flash->error(__('The user could not be deleted. Please, try again.'));
+        
+        if($user and $user['mail'] !== 'admin@admin.admin')
+        {
+            if($this->Users->delete($user)) {
+                $this->Flash->success(__('The user has been deleted.'));
+            } else {
+                $this->Flash->error(__('The user could not be deleted. Please, try again.'));
+            }
         }
 
-        return $this->redirect(['action' => 'index']);
+        else
+        {
+            $this->Flash->error(__('You just CANNOT delete the admin user of the website...'));
+        }
+
+        return $this->redirect('/');
     }
 
     public function login()
